--- test/integration/fn_cluster_ssh.go
@@ -1,35 +1,0 @@
-// +build integration
-
-/*
-Copyright 2016 The Kubernetes Authors All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package integration
-
-import (
-	"strings"
-	"testing"
-)
-
-func testClusterSSH(t *testing.T) {
-	MaybeParallel(t)
-	profile := Profile(t.Name())
-	mk := NewMinikubeRunner(t, profile, "--wait=false")
-	expectedStr := "hello"
-	sshCmdOutput, stderr := mk.MustRun("ssh echo " + expectedStr)
-	if !strings.Contains(sshCmdOutput, expectedStr) {
-		t.Fatalf("ExpectedStr sshCmdOutput to be: %s. Output was: %s Stderr: %s", expectedStr, sshCmdOutput, stderr)
-	}
-}
